modify readme
