--- sample_code_4/data/file1_in
@@ -1,3 +1,0 @@
-GEFEWFEWFGREGGEWG
-EREWRERREWRE
-
