--- sample_code_4/ksocket/Makefile
@@ -1,6 +1,0 @@
-obj-m += ksocket.o
-
-all:
-	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) modules
-clean:
-	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) clean
