--- sample_code_4/user_program/Makefile
@@ -1,5 +1,0 @@
-all:
-	gcc master.c -o master
-	gcc slave.c -o slave
-clean:
-	rm slave master
