+++ src/clean.sh
@@ -1,14 +1,12 @@
-cd ./ksocket
+cd ./master_device
 make clean
-rmmod ksocket.ko
-cd ../master_device
-make clean
-rmmod master_device.ko
+sudo rmmod master_device.ko
 cd ../slave_device
 make clean
-rmmod slave_device.ko
+sudo rmmod slave_device.ko
+cd ../ksocket
+make clean
+sudo rmmod ksocket.ko
 cd ../user_program
 make clean
 cd ..
-
-
