+++ src/compile.sh
@@ -1,12 +1,12 @@
 cd ./ksocket
 make
-insmod ksocket.ko
+sudo insmod ksocket.ko
 cd ../master_device
 make
-insmod master_device.ko
+sudo insmod master_device.ko
 cd ../slave_device
 make
-insmod slave_device.ko
+sudo insmod slave_device.ko
 cd ../user_program
 make
