--- src/ksocket/modules.order
@@ -1,1 +1,0 @@
-kernel//home/chj/course/os/project2/PJ2/3/modify/sample_light/ksocket/ksocket.ko
