+++ src/slave_device/Makefile
@@ -1,13 +1,7 @@
-MODULENAME = slave_device
+obj-m += slave_device.o
 
-obj-m += $(MODULENAME).o
-CFLAGS_$(MODULENAME) := -Wall
-
-all: module
-
-module:
-        make -C /lib/modules/$(shell uname -r)/build M=$(PWD) modules
-
+all:
+	cp ../ksocket/Module.symvers ./Module.symvers
+	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) modules
 clean:
-        make -C /lib/modules/$(shell uname -r)/build M=$(PWD) clean
-
+	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) clean
