+++ src/user_program/slave.c
@@ -28,26 +28,29 @@ int main (int argc, char* argv[])
 	strcpy(file_name, argv[1]);
 	strcpy(method, argv[2]);
 	strcpy(ip, argv[3]);
-
+	
 	if( (dev_fd = open("/dev/slave_device", O_RDWR)) < 0)//should be O_RDWR for PROT_WRITE when mmap()
 	{
 		perror("failed to open /dev/slave_device\n");
 		return 1;
 	}
+	
 	gettimeofday(&start ,NULL);
+	
 	if( (file_fd = open (file_name, O_RDWR | O_CREAT | O_TRUNC)) < 0)
 	{
 		perror("failed to open input file\n");
 		return 1;
 	}
-
+	
 	if(ioctl(dev_fd, 0x12345677, ip) == -1)	//0x12345677 : connect to master in the device
 	{
 		perror("ioclt create slave socket error\n");
 		return 1;
 	}
-
-
+	
+    write(1, "ioctl success\n", 14);
+    
 	switch(method[0])
 	{
 		case 'f'://fcntl : read()/write()
@@ -57,6 +60,8 @@ int main (int argc, char* argv[])
 				write(file_fd, buf, ret); //write to the input file
 				file_size += ret;
 			}while(ret > 0);
+			break;
+		case 'm': //mmap : magic!
 			break;
 	}
 
@@ -69,7 +74,7 @@ int main (int argc, char* argv[])
 	}
 	gettimeofday(&end, NULL);
 	trans_time = (end.tv_sec - start.tv_sec)*1000 + (end.tv_usec - start.tv_usec)*0.0001;
-	printf("Transmission time: %lf ms, File size: %d bytes\n", trans_time, file_size / 8);
+	printf("Slave transmission time: %lf ms, File size: %d bytes with method %s\n", trans_time, file_size / 8, method);
 
 
 	close(file_fd);
